Fix: CURAND pointer made 64-bit Fix: CURAND 32-bit support removed
git-svn-id: https://cudafy.svn.codeplex.com/svn@29302 2b2adeab-ffde-484a-ad65-8eada40eade2

+++ Cudafy/Cudafy.Math/RAND/CURandTypedefs.cs
@@ -59,7 +59,7 @@ namespace Cudafy.Maths.RAND
     [StructLayout(LayoutKind.Sequential)]
     internal struct RandGenerator
     {
-        public uint handle;
+        public ulong handle;
     }
 
     //public class RandGeneratorDevice : RandGenerator
